No longer hand wand image objects around.
The converter should not return any wand Image objects but only the
resulting scale blob with metadata.

The wand Image objects should only be used while converting images and
their scales. We do not want to store the full Image object, neither
in RAM nor in the ZODB.

The converter now returns a simple dictionary with the blob data and
metadata such as height, width, etc.